Moved static methods from NumberValue
