--- community/values/src/main/java/org/neo4j/values/NumberArrayValues.java
@@ -1,78 +1,0 @@
-/*
- * Copyright (c) 2002-2017 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * This file is part of Neo4j.
- *
- * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.neo4j.values;
-
-public class NumberArrayValues
-{
-    private NumberArrayValues()
-    {
-    }
-
-    public static boolean numbersEqual( IntegralArrayValue lhs, IntegralArrayValue rhs )
-    {
-        int length = lhs.length();
-        if ( length != rhs.length() )
-        {
-            return false;
-        }
-        for ( int i = 0; i < length; i++ )
-        {
-            if ( lhs.longValue( i ) != rhs.longValue( i ) )
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static boolean numbersEqual( FloatingPointArrayValue lhs, FloatingPointArrayValue rhs )
-    {
-        int length = lhs.length();
-        if ( length != rhs.length() )
-        {
-            return false;
-        }
-        for ( int i = 0; i < length; i++ )
-        {
-            if ( lhs.doubleValue( i ) != rhs.doubleValue( i ) )
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static boolean numbersEqual( FloatingPointArrayValue fps, IntegralArrayValue ins )
-    {
-        int length = ins.length();
-        if ( length != fps.length() )
-        {
-            return false;
-        }
-        for ( int i = 0; i < length; i++ )
-        {
-            if ( !NumberValue.numbersEqual( fps.doubleValue( i ), ins.longValue( i ) ) )
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-}
