+++ community/values/src/main/java/org/neo4j/values/NumberValue.java
@@ -19,106 +19,7 @@
  */
 package org.neo4j.values;
 
-import java.math.BigDecimal;
-
 public abstract class NumberValue extends ScalarValue
 {
-    private static final long NON_DOUBLE_LONG = 0xFFE0_0000_0000_0000L; // doubles are exact integers up to 53 bits
-
     abstract double doubleValue();
-
-    // TODO: this is a lousy hash.
-    public static int hash( long number )
-    {
-        return (int) (number ^ (number >>> 32));
-    }
-
-    public static int hash( double number )
-    {
-        long asLong = (long) number;
-        if ( asLong == number )
-        {
-            return hash( asLong );
-        }
-        long bits = Double.doubleToLongBits( number );
-        return (int)(bits ^ (bits >>> 32));
-    }
-
-    public static boolean numbersEqual( double fpn, long in )
-    {
-        if ( in < 0 )
-        {
-            if ( fpn < 0.0 )
-            {
-                if ( (NON_DOUBLE_LONG & in) == NON_DOUBLE_LONG ) // the high order bits are only sign bits
-                { // no loss of precision if converting the long to a double, so it's safe to compare as double
-                    return fpn == in;
-                }
-                else if ( fpn < Long.MIN_VALUE )
-                { // the double is too big to fit in a long, they cannot be equal
-                    return false;
-                }
-                else if ( (fpn == Math.floor( fpn )) && !Double.isInfinite( fpn ) ) // no decimals
-                { // safe to compare as long
-                    return in == (long) fpn;
-                }
-            }
-        }
-        else
-        {
-            if ( !(fpn < 0.0) )
-            {
-                if ( (NON_DOUBLE_LONG & in) == 0 ) // the high order bits are only sign bits
-                { // no loss of precision if converting the long to a double, so it's safe to compare as double
-                    return fpn == in;
-                }
-                else if ( fpn > Long.MAX_VALUE )
-                { // the double is too big to fit in a long, they cannot be equal
-                    return false;
-                }
-                else if ( (fpn == Math.floor( fpn )) && !Double.isInfinite( fpn ) )  // no decimals
-                { // safe to compare as long
-                    return in == (long) fpn;
-                }
-            }
-        }
-        return false;
-    }
-
-    // Tested by PropertyValueComparisonTest
-    public static int compareDoubleAgainstLong( double lhs, long rhs )
-    {
-        if  ( (NON_DOUBLE_LONG & rhs ) != NON_DOUBLE_LONG )
-        {
-            if ( Double.isNaN( lhs ) )
-            {
-                return +1;
-            }
-            if ( Double.isInfinite( lhs ) )
-            {
-                return lhs < 0 ? -1 : +1;
-            }
-            return BigDecimal.valueOf( lhs ).compareTo( BigDecimal.valueOf( rhs ) );
-        }
-        return Double.compare( lhs, rhs );
-    }
-
-    // Tested by PropertyValueComparisonTest
-    public static int compareLongAgainstDouble( long lhs, double rhs )
-    {
-        return - compareDoubleAgainstLong( rhs, lhs );
-    }
-
-    /**
-     * Return an integer one less than the given integer, or throw {@link ArithmeticException} if the given integer is
-     * zero.
-     */
-    public static int decrementExactNotPastZero( int value )
-    {
-        if ( value == 0 )
-        {
-            throw new ArithmeticException( "integer underflow past zero" );
-        }
-        return value - 1;
-    }
 }
